feat(index.js): Add safe update/create functions
This change introduces a new set of `write*(version, name, data)` functions such as `writeS3()` or
`writeVCL` that allow the safe creation or update of resources. Calling these methods will create
the resource, and if it already exists, will update it. This reduces the amount of error handling
needed in consumer code.