docs(readme): add deprecation warning
development on the official `fastly` NPM package has started again, there will be a modern
ES6-module based implementation that covers all of the Fastly API and is maintained by the
Fastly team, so there is no point in keeping this package around for any longer. I'm adding
a deprecation warning to the readme, indicating that a deprecation is likely imminent.